+++ lib/figjam/application.rb
@@ -5,6 +5,10 @@ require "figjam/figaro_alias"
 module Figjam
   class Application
     FIGARO_ENV_PREFIX = "_FIGARO_"
+    SILENCE_STRING_WARNINGS_KEYS = %i[
+      FIGARO_SILENCE_STRING_WARNINGS
+      FIGJAM_SILENCE_STRING_WARNINGS
+    ]
 
     include Enumerable
 
@@ -70,8 +74,10 @@ module Figjam
     end
 
     def set(key, value)
-      non_string_configuration!(key) unless key.is_a?(String)
-      non_string_configuration!(value) unless value.is_a?(String) || value.nil?
+      unless non_string_warnings_silenced?
+        non_string_configuration!(key) unless key.is_a?(String)
+        non_string_configuration!(value) unless value.is_a?(String) || value.nil?
+      end
 
       ::ENV[key.to_s] = value.nil? ? nil : value.to_s
       ::ENV[FIGARO_ENV_PREFIX + key.to_s] = value.nil? ? nil: value.to_s
@@ -79,6 +85,13 @@ module Figjam
 
     def skip?(key)
       ::ENV.key?(key.to_s) && !::ENV.key?(FIGARO_ENV_PREFIX + key.to_s)
+    end
+
+    def non_string_warnings_silenced?
+      SILENCE_STRING_WARNINGS_KEYS.any? { |key|
+        # Allow the silence configuration itself to use non-string keys/values.
+        configuration.values_at(key.to_s, key).any? { |cv| cv.to_s == 'true' }
+      }
     end
 
     def non_string_configuration!(value)